+++ packages/example/lib/vision_detector_views/label_detector_view.dart
@@ -78,8 +78,8 @@ class _ImageLabelViewState extends State<ImageLabelView> {
       _text = '';
     });
     final labels = await _imageLabeler.processImage(inputImage);
-    if (inputImage.inputImageData?.size != null &&
-        inputImage.inputImageData?.imageRotation != null) {
+    if (inputImage.metadata?.size != null &&
+        inputImage.metadata?.rotation != null) {
       final painter = LabelDetectorPainter(labels);
       _customPaint = CustomPaint(painter: painter);
     } else {