+++ packages/google_mlkit_commons/ios/Classes/MLKVisionImage+FlutterPlugin.m
@@ -27,40 +27,16 @@
 + (MLKVisionImage *)bytesToVisionImage:(NSDictionary *)imageData {
     FlutterStandardTypedData *byteData = imageData[@"bytes"];
     NSData *imageBytes = byteData.data;
-    
     NSDictionary *metadata = imageData[@"metadata"];
-    NSArray *planeData = metadata[@"planeData"];
-    size_t planeCount = planeData.count;
-    
     NSNumber *width = metadata[@"width"];
     NSNumber *height = metadata[@"height"];
-    
-    NSNumber *rawFormat = metadata[@"imageFormat"];
-    FourCharCode format = FOUR_CHAR_CODE(rawFormat.unsignedIntValue);
-    
-    CVPixelBufferRef pxBuffer = NULL;
-    if (planeCount == 0) {
-        @throw [NSException exceptionWithName:NSInvalidArgumentException
-                                       reason:@"Can't create image buffer with 0 planes."
-                                     userInfo:nil];
-    } else if (planeCount == 1) {
-        NSDictionary *plane = planeData[0];
-        NSNumber *bytesPerRow = plane[@"bytesPerRow"];
-        pxBuffer = [self bytesToPixelBuffer:width.unsignedLongValue
-                                     height:height.unsignedLongValue
-                                     format:format
-                                baseAddress:(void *)imageBytes.bytes
-                                bytesPerRow:bytesPerRow.unsignedLongValue];
-    } else {
-        pxBuffer = [self planarBytesToPixelBuffer:width.unsignedLongValue
-                                           height:height.unsignedLongValue
-                                           format:format
-                                      baseAddress:(void *)imageBytes.bytes
-                                         dataSize:imageBytes.length
-                                       planeCount:planeCount
-                                        planeData:planeData];
-    }
-    
+    NSNumber *rawFormat = metadata[@"image_format"];
+    NSNumber *bytesPerRow = metadata[@"bytes_per_row"];
+    CVPixelBufferRef pxBuffer = [self bytesToPixelBuffer:width.unsignedLongValue
+                                                  height:height.unsignedLongValue
+                                                  format:FOUR_CHAR_CODE(rawFormat.unsignedIntValue)
+                                             baseAddress:(void *)imageBytes.bytes
+                                             bytesPerRow:bytesPerRow.unsignedLongValue];
     return [self pixelBufferToVisionImage:pxBuffer];
 }
 
@@ -72,47 +48,6 @@
     CVPixelBufferRef pxBuffer = NULL;
     CVPixelBufferCreateWithBytes(kCFAllocatorDefault, width, height, format, baseAddress, bytesPerRow,
                                  NULL, NULL, NULL, &pxBuffer);
-    return pxBuffer;
-}
-
-+ (CVPixelBufferRef)planarBytesToPixelBuffer:(size_t)width
-                                      height:(size_t)height
-                                      format:(FourCharCode)format
-                                 baseAddress:(void *)baseAddress
-                                    dataSize:(size_t)dataSize
-                                  planeCount:(size_t)planeCount
-                                   planeData:(NSArray *)planeData {
-    size_t widths[planeCount];
-    size_t heights[planeCount];
-    size_t bytesPerRows[planeCount];
-    
-    void *baseAddresses[planeCount];
-    baseAddresses[0] = baseAddress;
-    
-    size_t lastAddressIndex = 0;  // Used to get base address for each plane
-    for (int i = 0; i < planeCount; i++) {
-        NSDictionary *plane = planeData[i];
-        
-        NSNumber *width = plane[@"width"];
-        NSNumber *height = plane[@"height"];
-        NSNumber *bytesPerRow = plane[@"bytesPerRow"];
-        
-        widths[i] = width.unsignedLongValue;
-        heights[i] = height.unsignedLongValue;
-        bytesPerRows[i] = bytesPerRow.unsignedLongValue;
-        
-        if (i > 0) {
-            size_t addressIndex = lastAddressIndex + heights[i - 1] * bytesPerRows[i - 1];
-            baseAddresses[i] = baseAddress + addressIndex;
-            lastAddressIndex = addressIndex;
-        }
-    }
-    
-    CVPixelBufferRef pxBuffer = NULL;
-    CVPixelBufferCreateWithPlanarBytes(kCFAllocatorDefault, width, height, format, NULL, dataSize,
-                                       planeCount, baseAddresses, widths, heights, bytesPerRows, NULL,
-                                       NULL, NULL, &pxBuffer);
-    
     return pxBuffer;
 }