+++ packages/google_mlkit_commons/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.4.0
+
+* Update InputImage.
+* Rename InputImageData -> InputImageMetadata.
+* Remove InputImagePlaneMetadata.
+* Only support nv21 for Android and bgra8888 for iOS.
+
 ## 0.3.0
 
 * Update dependencies.